requirements.txt
streamlit >= 1.37
facebook-business>=19.0.6
python-dotenv>=1.0.1

streamlit
requests

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery

facebook-business
streamlit-extras